Show both custom and default menu items when selection is present.
BUG=none
TEST=none

We used to omit default context menu items if custom items are present.
Now we include both custom and default items in case there's an active selection.

We need this to be able to show custom context menu items on selection in DevTools. Here's the WebKit part of the change:

https://bugs.webkit.org/show_bug.cgi?id=69924

Review URL: http://codereview.chromium.org/8318005