Use new triggers for all reliability bots.
The current method of triggering the reliability
bots causes the changeset data not to be sent to the
reliability test run and appear with the rest of the
build data on the build waterfall.

Switch to the new triggers that are configured based on
the factory properties settings which are used pretty
much everywhere else in Chromium's Buildbot and do send
the changeset data to triggered bots.

BUG=98045


Review URL: http://codereview.chromium.org/8373017