Revert 107360
This broke the Linux Touch bot

This patch lets the browser window punch a hole in the DesktopBackgroundView,
thus eliminating overdraw.
To do this, the hole punching logic was changed to look at the entire tree.
Drawing the screen is a preorder traversal through all the layers. Thus, to
compute the hole, we have to look at all of the layers after a particular layer
in the preorder traversal.
This is not an efficient algorithm, but it is simplest I could come up with.
Plus we will be moving to the WebKit compositor soon.

Hopefully this will improve the frame rate for the Aura demos

Note: I will put the changes to Transform into another patch.

Review URL: http://codereview.chromium.org/8368013

TBR=pkotwicz@chromium.org
Review URL: http://codereview.chromium.org/8394054