Fix generation of API documentation for arrays of objects
If a type reflects an array of objects with specified properties, such as here
  "id": "HttpHeaders",
  "type": "array",
  "items": {
    "type": "object",
    "properties": {...}
  },
the API documentation generator would swallow the contents of "properties". The documentation would just contain the information that an array of objects is passed, without details of the keys stored in the objects. This is fixed in this CL.


BUG=no
TEST=no


Review URL: http://codereview.chromium.org/8411001