aura: Draw persistent borders around browser windows.
This makes BrowserNonClientFrameViewAura's translucent gray
background always stick out a bit beyond the sides and
bottom of windows, instead of just peeking out when the user
hovers for resizing.

I originally added a separate view for this, but it looked
strange since Chrome already draws a shadow along the top of
the tab strip.

BUG=101977
TEST=manual

Review URL: http://codereview.chromium.org/8343066