aura: Work around RenderWidgetHostTest.Background segfault.
This test was creating an aura::Window but not calling its
Init() method, resulting in a segfault when
RenderWidgetHostViewAura::SetBackground() tried to access
the window's layer.

I have a more ambitious change in progress to add a
RenderWidgetHostView::InitAsChild() method and update a
bunch of callers; this is just a quick workaround so try
bots will stop failing.

BUG=102450
TEST=manual: test no longer segfaults

Review URL: http://codereview.chromium.org/8432020