Add support for temporarily disabling sync.
Adds functionality to the ProfileSyncService to temporarily disable
sync and later be able to enable it again.

This is achieved by stopping the sync backend and suppressing future
startups when sync is disabled. The flag for suppressing startup is
removed when enabling sync again, and the sync backend is started.

Contributed by nyquist@chromium.org

TEST=unit_tests


Review URL: http://codereview.chromium.org/8332023