SunSpider benchmark test in perf.py now outputs results for each category.
The SunSpider suite outputs a total overall result, followed by a set of
categories, where each category includes a total result for that category,
along with the individual benchmark results within that category.
Previously, the SunSpider test in perf.py did not output the result for
each category to be graphed.  It now does.

BUG=chromium-os:18457
TEST=None


Review URL: http://codereview.chromium.org/8386048