Aura: prevent synchronous drawing of web content from triggering an extra SchedulePaint
When resizing, we try to get the renderer to paint synchronously within a small
delay. When that happens, RWHVA::OnPaint calls RWH::GetBackingStore which calls
back RWHVA::DidUpdateBackingStore. We don't want that to schedule a paint
(including a compositing pass) for later, since we're already about to resolve
the resulting damage.

BUG=None
TEST=


Review URL: http://codereview.chromium.org/8386054