Add a test suite to PYAUTO_TESTS for perf tests for Seaboard
I am not sure if this is the right way of doing this.
The goal of the CL is to add a test suite where we can gradually add the perf tests for seaboard one by one.
Currently, running the full suite crashes the device.

Review URL: http://codereview.chromium.org/8432036