Add a UMA histogram to count accessibility usage.
Deletes code that incorrectly used a stat counter instead of a histogram,
and adds a function that runs 10 seconds after the browser starts and
updates a UMA enum histogram with the status of whether accessibility is
enabled or not. Also makes it so that manually enabling accessibility
from the command-line triggers this as well - we may want to separate
this out later.

BUG=99504
TEST=none
Review URL: http://codereview.chromium.org/8479015