Fix bug where wglGetExtensionsStringEXT was being called instead of wglGetExtensionsStringARB.
This crashed because they have different signatures.

I switched to always use wglGetExtensionsStringEXT because it avoids the need to get the DC of a surface which was always a bit awkward.

Also fixed a bug where WGL contexts were not put in the right share group.
Review URL: http://codereview.chromium.org/8481012