Change EXTENSION_PROCESS_CREATED observers to use EXTENSION_HOST_CREATED.
Change EXTENSION_HOST_CREATED to be fired on RenderViewReady.
Remove the EXTENSION_PROCESS_CREATED notification.

This is related to fixing bug 71629, in particular, making the correct things appear in the task manager when the extension host reuses a process. This affects the new reload path (but only for unpacked extensions), but it's also visible when a process isn't created for a popup in lots-of-tabs situations.

The extension settings page (the other observer) already listens to enough notifications that it updates properly (too many of them, but that's another story -- see bug).

BUG=102950
TEST=TaskManagerBrowserTest.* and also: Open Task Manager. Open an extension popup, extension notification, and extension page in a tab; check that they all appear. Kill an extension with a background page and reload it from the notification bubble; check that it reappears.

Review URL: http://codereview.chromium.org/8480007