Use WebUI login screen to authenticate on WebUI lock screen.
This CL uses the existing WebUI login screen to let the user authenticate for the WebUI lock screen. In a future CL changes will be made to the WebUI login screen in the case where a user is logged in (screen is locked) to match the new lock screen UI mockups.

BUG=100718
TEST=Enable WebUI lock screen in chrome://flags, lock the screen and authenticate with login page to test.


Review URL: http://codereview.chromium.org/8403004