Move flush call after resource deletion into SharedIdHandler
The SharedIdHandler is the only id allocator that makes client ids
available to other contexts in the share group, so it's the only one
that needs a Flush() call on resource deletion. Additionally it only
needs a command buffer level flush to ensure execution order and not a
glFlush().

BUG=none

Review URL: http://codereview.chromium.org/8508043