Add support for --critical-update-version to installer.
The installer may now be given a --critical-update-version=W.X.Y.Z option on the command line.  If this is present for an in-use update and the indicated version is newer than the in-use Chrome, a new "cpv" value is dropped in the registry alongside "opv" and "cmd" (the other state associated with an in-use update).  Thanks to Finnur, Chrome will eventually notice that there's a pending critical update and do magical things to keep our users safe and secure.  Go users!

Other things I did while I was at it:
- switched version string conversions from UTF8ToWide to ASCIIToWide since version numbers are always ASCII dotted numbers.
- renamed the reg value from "CriticalUpdate" to "cpv" so it fits in nicely with the other cryptic values in the product's Clients key (a.k.a. Version key).

BUG=103526,97665
TEST=Install version N, launch it with --check-for-update-interval=5, then install verison M>N with --critical-update-version=M.  After a few seconds, Chrome should do something fantastic.  Note that the alternate_version_generator can be used to create a newer versioned mini_installer from an existing one (just build it and run it).


Review URL: http://codereview.chromium.org/8517012