Convert render thread idle delays from seconds to milliseconds.
This CL eliminates double to integer conversions when scheduling the idle timer and allows subsecond delays.

The existing code simply rounds the delay down to an integer when scheduling the idle timer, which means that any subsecond delay schedules a timer with zero delay.

BUG=
TEST=


Review URL: http://codereview.chromium.org/8463019