Set the profiler's thread name at a central place
This will automatically catch thread names in several of the
child processes we are after, and it also catches the name
for a webkcore thread.

The underlying function uses a try/catch clause, which
contaminates the entire function.  As a result, I needed
to factor out that section, so that I could still call
to set the profiler thread name.  Without this, the
shared dbg build will fail to compile.

r=rtenneti
Review URL: http://codereview.chromium.org/8565036