Fix Panel to always detect when a tab is added.
This solves the problem when chrome.windows.create() is used to create a panel with a tabid. Panel was watching for NOTIFICATION_TAB_ADDED, which is only sent upon browser navigation. As there is no navigation when existing tab contents are used, the Panel does not set up auto-resize in this scenario.

BUG=None
TEST=PanelBrowserTest.CreateWithExistingContents

Review URL: http://codereview.chromium.org/8539025