Reland http://codereview.chromium.org/8477015 Landed at r109114, Reverted at r109142
Make it possible to enable/disable specific ppapi tests. Migrate PostMessage tests.

The testcase attribute now can include a 'filter'. If it's omitted, everything works the same as before. This way we can migrate tests over bit-by-bit if we want to. We can also still run the tests manually the same way as before.

This only runs PostMessage testss the new way, and re-enables all oop PostMessage tests that pass on Windows. I can do the other tests in this CL if desired, but it might be easier to land in a few pieces.

BUG=102885
TEST=N/A
TBR=dmichael@chromium.org

Review URL: http://codereview.chromium.org/8536031