Fix for management API related to escalated permissions disabled extensions
The general approach is to require user gesture and put up the confirmation UI if an extension wants to re-enable a disabled-due-to-permissions-escalation extension.

Also added some new infrastructure for making this easier to test.

(This is re-landing http://codereview.chromium.org/8423069/ which had a compile
 error on the cros_tegra2 bot)

BUG=102579
TEST=See bug

TBR=aa@chromium.org

Review URL: http://codereview.chromium.org/8488012