Avoid too many consecutive redraws on about:profiler while loading.
Background: when first opening about:profiler, the browser sends data for each process as it becomes available. The UI responds to this by redrawing all the data. If there are a lot of processes then this can cause many conecutive repaints which slows things down. To avoid this, we now bulk together redraws on a 500ms timer (just for loading, not any user-initiated action).

BUG=100992
Review URL: http://codereview.chromium.org/8585023