Added some debugging code to print layer tree.
Also added a name attribute to the layer. IF !NDEBUG is defined, layer takes on the name of the view/window which created it

Please comment as to where the debug_utils.cc should be located, it feels like the wrong place.

I have another patch to print the window tree which I'll submit for review once I get this one in.

I think this is a bit simpler and more useful than http://codereview.chromium.org/8366008/

Review URL: http://codereview.chromium.org/8508009