Replace FileManagerUtil class with namespace file_manager_util.
file_manager_util namespace is consistent with other utils in the codebase (platform_util, mime_util, file_util), and is the preferred way in the google style guide to do this kind of thing. The FileManagerUtil class is never initialized and has no state.

BUG=None, cleanup
TEST=Compiles, tests pass.
Review URL: http://codereview.chromium.org/8574054