Prepare Chromium for WebCompositor::initialize.
Currently, we have two duplicate mechanisms for telling the compositor to go into threaded mode. First, you set WebCompositor::setThread, then you pass enableCompositorThread via WebPreferences.

We're moving to a global initialize/shutdown flow on the compositor. This is a breaking change on the WebKit API so this change prepares us for it.

Review URL: http://codereview.chromium.org/8558030