aura: Re-enable XInput2 events.
It turns out that the position information in an event for a slave device
reported by X is always slightly behind than the actual position. However, it
turns out that the valuators do contain the correct information. So get the
event location from the valuator.

BUG=103981
TEST=manually

Review URL: http://codereview.chromium.org/8568044