Update last used profile when deleting profile
Deleting a profile can leave the user in a state where their last used profile pionts to a deleted profile. This can lead to crashes when the user opens a new window or uses the menu bar.

Fix was to update the last used profile when deleting a profile.

BUG=104957
TEST=Created 3 profiles, with just one window open deleted the active profile. Verified that File -> New Window opened a window to the next profile. Verified that Users -> Customize customized the next profile.


Review URL: http://codereview.chromium.org/8592006