Make Aura and compositor stacking methods more intuitive.
aura::Window::StackChildAbove() and ui::Layer::StackAbove()
stacked entity A directly above entity B... unless A was
already somewhere above B in the stack, in which case they'd
do nothing.  I see the caveat as making these methods less
useful, and I'm unable to think of any benefits from it, so
this change makes the methods place A directly above B in
all cases (possibly lowering A in the stacking order if it
was previously far above B).

This also fixes an issue in the Aura shell's shadow-stacking
code where ShadowController could inadvertently raise a
window to the top of the stacking order.

BUG=101977
TEST=added

Review URL: http://codereview.chromium.org/8653005