Refactor ContinueWindow::Show() to accept a callback parameter.
This gets rid of the DesktopEnvironment::OnPause() notification from the Host.
Instead, that code is moved to DesktopEnvironment::ContinueSession() and
provided as a callback function parameter to ContinueWindow::Show().

This will simplify a future CL to move the ContinueWindow/DisconnectWindow code
into an IT2Me-specific HostObserver subclass.


BUG=None
TEST=No change in functionality.


Review URL: http://codereview.chromium.org/8624009