Don't update jabberId when renaming a host.
Previous the webapp specified jabberId in the update request when a host
is rename. To the directory this looks as updated jabberId, and so the
status is set to ONLINE. Another problem was that update request was
also setting the status field, but it shouldn't.

Review URL: http://codereview.chromium.org/8679017