Add custom-appearance to sync sign in button
The sync sign in button has a custom gaia look. When disabled though, the button became unreadable.

The problem was due to a change I made to button.css in r109201. I allowed buttons to opt out of the standard button look by specifying class="custom-appearance". This caused the look of the GAIA button to change.

To fix this I'm setting class="custom-appearance" to the GAIA button as well.

BUG=104176
TEST=Ran without my fix and saw that the disabled sign in button was unreadble. Verified that with my fix the disabled sign in button looked correct.

Review URL: http://codereview.chromium.org/8713005