Cocoa: Fix avatar menu shortcut
If focus was in the Omnibox then pressing Command-M would incorrectly trigger the avatar menu.

The problem was due to the fact that when MatchesEventForKeyboardShortcut() matches against a key character it ignores the shift flag.

Fix was to use a key code for the shortcut so that the shift flag was correctly checked.

BUG=104457
TEST=Verified that Command-Shift-M correctly showed the avatar menu while focused on a web page and while focused on the location bar. Verified that Command-M correctly minimized the window while focused on a web page and focused on the location bar.

Review URL: http://codereview.chromium.org/8718002