Enable vsync on CompositorCC.
This is relying on the fact that the WebGraphicsContext3DInProcessImpl has a gfx::GLContext under the hood. The alternative is to push SetSwapInterval into the initialization of WebGraphicsContext3DInProcessImpl. I've preferred to do it here, because I think that they other clients of WebGraphicsContext3DInProcessImpl don't need vsync enabled.

BUG=none
TEST=Build chrome using GYP_DEFINES="use_aura=1 use_webkit_compositor=1". Run chrome with proprietary Nvidia driver (supports vsync). See no tearing when dragging the window around.

Review URL: http://codereview.chromium.org/8700007