Move the comparators for QuotaInfo into the class.
My version of MSVC 2008 has a bug in ADL that prevents it from finding the
operators in the global namespace if the class in question is nested.  This
meant I was unable to compile the BrowsingDataQuotaHelperTest.FetchData test
because it tried to use std::set<QuotaInfo>.  Moving the operators into the
class itself solves the problem, and should work properly with all other
compilers as well.

TEST=BrowsingDataQuotaHelperTest
R=tzik@chromium.org

Review URL: http://codereview.chromium.org/8735001