aura: Turn XInput2 back on.
The previous attempt read values from the valuator without properly checking if
the valuator is set or not. This resulted in incorrect location at times (when X
or Y didn't change). So this change makes sure the valuator is set correctly
before reading from it.

BUG=103981
TEST=manually (on desktop and device)

Review URL: http://codereview.chromium.org/8634009