Replace \n in message strings to acutal new lines.
I had added a hack in http://codereview.chromium.org/196007 to mimic a
behavior of how rc.exe replaces \n with a newline.  Unfortunately, this
didn't make it possible to have legitimate uses of \n in a string.

Instead, remove all occurrences of \n in preparation of removing this hack.

BUG=105985

Review URL: http://codereview.chromium.org/8757009