Improve Autocomplete Matches and Handling of Large Results Sets
Do not call FixupUserInput as it was prepending unexpected prefixes (such as file://) to the search string and bypassing valid results.
Move the search string decomposition operation from the HQP into the IMUI.
In the final substring filtering use whitespace delineated terms rather than words.
Instead of bailing if we get a large results set (>500) filter it down to 500 by sorting by typed-count/visit-count/last-visit.
This means it's no longer necessary to bypass the HQP if there is only one character in the search term so get rid of the ExpandedInMemoryURLIndexTest.ShortCircuit unit test.

BUG=101301,103575
TEST=Added unit tests.
Review URL: http://codereview.chromium.org/8526010