Remove IOThread::message_loop().
Instead introduce BrowserThread::UnsafeGetMessageLoop which is used in
a couple of places.  This will be removed in a follow-up change.  I
could have used UnsafeGetBrowserThread(id)->message_loop() but this is
even easier to search for.

Also, remove a couple of BrowserProcess accessors that aren't used by anybody, and that were using UnsafeGetBrowserThread.

TBR=battre@chromium.org
BUG=98716


Review URL: http://codereview.chromium.org/8775014