Revert 113027 - Zygote most of the uses of the utility process on Linux
Allow the user of the utility process to specify whether the zygote
should be used on Linux. This allows to sandbox all the uses
of the utility process that don't do FS access, which right now are all
except extension unpacking and NPAPI plugin listing.

This is the first step to get the utility process sandboxed 
on Linux. Since most of the uses of the utility process 
don't do file access, launching all of those from the zygote 
will simplify sandboxing the one that does: extension 
unpacking.

BUG=93109
TEST=Try installing an extension from the web store.


Review URL: http://codereview.chromium.org/8770025

TBR=jorgelo@chromium.org
Review URL: http://codereview.chromium.org/8817013