Chrome Frame test fixes for Vista/IE7.
Write the ActiveX control from an external script in all test documents.  This works around a "feature" of IE versions released between April, 2006 and April, 2008 that required the user to click on an ActiveX control to "activate" it.  See http://msdn.microsoft.com/en-us/library/ms537508.aspx.

I also added asserts for the copying of CFInstance and CFInstall so that the whole test aborts if the files can't be copied, since otherwise test results are bogus.

BUG=60987
TEST=no manual testing needed; chrome_frame_tests.exe will either become better or worse.

Review URL: http://codereview.chromium.org/8777002