Revert of 112134 of Revert 112130 - Close idle connections / SPDY sessions when needed.
Due to the idle connection state being held by different socket pools, it's possible for one socket pool to hold an idle socket in a lower layer socket pool. From the lower level socket pool's perspective, the socket is being "actively" used. From the higher socket pool's (including SpdySession, which is more of a connection manager) perspective, the connection is idle and can be closed if we have hit a limit.

Normally this isn't a big deal, except when we have a lot of idle SPDY connections and are connecting via a proxy, so we have low connection limits through the proxy server. We address this problem by allowing lower-level socket pools to tell higher level socket pools to close a socket. 

Fixed ASAN test failures by removing  .Times(1) and .Times(2)  from CloseMultipleIdleSocketsHeldByLayeredPoolWhenNeeded unittest (this removes the tests relying on the order of std::set in CloseOneIdleConnectionInLayeredPool). ASAN is probably causing the memory allocator to allocate the pools differently. The std::set is ordered by LayeredPool* which is the address of the LayeredPool (willchan).


BUG=62364,92244, 105839
TEST=none


Review URL: http://codereview.chromium.org/8340012

TBR=willchan@chromium.org
Review URL: http://codereview.chromium.org/8745007

TBR=willchan@chromium.org

Review URL: http://codereview.chromium.org/8803019