sync: change semantics (and name) of SigninManager::GetUsername
This is an incremental change rolled out of a larger one to remove cros_user, and make
the ProfileSyncService<Foo>Test framework easier to configure / decouple from the cros case. 
In the upcoming patch I'll remove cros_user and in several cases rely on SigninManager's new GetAuthenticatedUsername.

Also makes it so we always initialize the signin manager, regardless of platform.

Also fix a case in ProfileSyncService that calls DisableForUser when a UI error seems more appropriate.

This effectively re-implements the fix for bug 103861.

BUG=88109
TEST=


Review URL: http://codereview.chromium.org/8698001