Aura Shell singleton needs to be destroyed prior to the Aura Desktop
Adds code to explicitly destroy the Aura Shell class prior to the destruction of Aura Desktop.  The reason for this is that the Shell leaks the DesktopFilter event listeners otherwise and DCHECKs fire.

Note, this currently only affects the Mac port because other platforms exit() from the run loop and don't do explicit tear-down.

BUG=none
TEST=Manually run aura_shell_exe on Mac.
R=sky@chromium.org, xiyuan@chromium.org

Review URL: http://codereview.chromium.org/8824017