Fix loss of mouse events when an Aura combobox is open in a modal dialog.
NativeComboboxViews was not passing itself as a (transient) parent to
MenuRunner::RunMenuAt. All mouse events were sent to the open menu (since it
is set as the capture window), and ModalityEventFilter::PreHandleMouseEvent
dropped them all since they were not directed to a child of the active modal
dialog.


BUG=105155
TEST=none


Review URL: http://codereview.chromium.org/8893019