Update Java Bridge to use its own thread in the browser process
Currently the Java Bridge uses the WEBKIT thread to manipulate injected objects
in the browser process. However, the WEBKIT thread is not set up when the
renderer is run in-process, and it may be deprecated.

This change adds a new background thread in JavaBridgeDispatcherHost, and updates the JavaBridge to use it in place of the WEBKIT thread.

BUG=106838


Review URL: http://codereview.chromium.org/8890018