Merge 113926 - Fix cause of uninitialized memory read in ExtensionService unit_tests
In crrev.com/113111 I checked in code that reads a app state boolean from an
ExtensionSyncData object which is uninitialized. This fixes the initializati

for the relevant constructor, and also limits reading that value to only whe

the extension is an app.

BUG=106549
TEST=valgrind run of ExtensionService sync-related unit_tests should come back
clean
    
Review URL: http://codereview.chromium.org/8899010

TBR=asargent@chromium.org
Review URL: http://codereview.chromium.org/8922012