Fix a bug where the JavaBridgeDispatcherHost is prematurely deleted
JavaBridgeDispatcherHost must be ref-counted because it is used with
base::Bind. http://codereview.chromium.org/8834013 changed
JavaBridgeDispatcherHost to also inherit from RenderViewHostObserver. The base
implementation of RenderViewHostObserver::RenderViewHostDestroyed() deletes the
object. This is unwanted in this class, so we provide an empy implementation of
the method.


Review URL: http://codereview.chromium.org/8929005