Merge 114330 - Update avatar bubble if profile info changes
Showing the avatar menu for the first time with two or more GAIA photos would cause the menu to look forzen.

The problem was that we the menu wasn't being resized and repositioned.

Fix was to explicitly resize and reposition the bubble.

BUG=106989
TEST=Reproduced the bug. Applied my fix and verified that the menu updated correctly.


Review URL: http://codereview.chromium.org/8907001

TBR=sail@chromium.org
Review URL: http://codereview.chromium.org/8890092