Merge 114212 - Fix the ownership check for signed settings.
The check was trying to verify the current user is the owner when all it
needs to check is if there is an owner already.

This check was also blocking local owner from doing changes to the signed
settings on first login and possibly is the cause preventing the managed
settings from getting applied immediately too.

BUG=chromium-os:24020,chromium-os:24007
TEST=As described in the bugs.

Review URL: http://codereview.chromium.org/8921007

TBR=pastarmovj@chromium.org
Review URL: http://codereview.chromium.org/8907031