[Sync] Add HistoryService::DeleteURLsForTest
This is so that the sync performance tests don't need to wait for each
URL delete to hit the disk before moving on to the next one.

This is a speculative fix to stabilize the delete_typed_urls sync perf
test.  It also cuts down the runtime of the typed urls test from >20s to <10s or so.

BUG=107040
TEST=

Review URL: http://codereview.chromium.org/8890089