Breakpad: Compile Breakpad into Chromium by default on Mac
This is intended to make it easier to test Breakpad crash dumping on
the buildbots.  Crash dumping will be disabled in Chromium by default
at run time unless CHROME_HEADLESS is set.

The bots set CHROME_HEADLESS, but we don't really want the bots to be
uploading crash dumps that are useless without symbols, so uploading
is disabled except for official builds and when building with
mac_breakpad_uploads=1.

This should not affect stack backtraces that occur in tests (such as
browser_tests) via StackDumpSignalHandler().  Although Breakpad would
normally take priority over this signal handler, Breakpad does not get
enabled in these test suites because is_browser is false inside
InitCrashReporter().

This change leaves the Gyp option "mac_breakpad" for enabling the
creation of Breakpad files.  This remains off by default for Chromium
builds, because the symbol dumping is slow and because it has problems
in the makefile-based Mac builds.

BUG=105778
TEST=run Chromium with --crash-test and CHROME_HEADLESS=1; a *.dmp file
  should appear in ~/Library/Application Support/Chromium/Crash Reports


Review URL: http://codereview.chromium.org/8824003