Move content::MakeSharedMemorySegmentViaIPC into its own file
The nacl_helper program uses content::MakeSharedMemorySegmentViaIPC but
none of the other proxy functions in the same source file.  Some of those
functions refer to symbols that bring in a great deal of other code, e.g.
most of v8.  We don't want all that code linked into nacl_helper--it bloats
the binary by a factor of ten or so.

Moving the function out to its own file prevents the other dependencies
coming in unnecessarily.

BUG= http://code.google.com/p/chromium/issues/detail?id=106986
TEST= still builds and nacl_helper binary is much smaller

R=jam@chromium.org,bradchen@google.com

Review URL: http://codereview.chromium.org/8893009