Fix race condition in ChromeProxyConfigService
The ChromeProxyConfigService was initialized to CONFIG_UNSET (i.e. effectively
a fallback to the system configuration) until the UI thread reported the
preference stored in the PrefService. This created a race condition during
startup where restored tabs could send requests while the
ChromeProxyConfigService was not configured with the values from the
PrefService, yet.

BUG=107326
TEST=no


Review URL: http://codereview.chromium.org/8951012