Merge 114906 - Fix Chrome Frame crash when automation server dies.
GCF was crashing if the automation server dies while there are outstanding sub-resource requests being serviced on the new bg thread.  The reason was that AutomationServerDied posts a task to InitializeComplete(AUTOMATION_SERVER_CRASHED) and then a task to Uninitialize().  The former will call ReleaseAutomationServer() without first joining on the bg resource thread.  The fix is to wait and release the server in Uninitialize.

BUG=107811
TEST=Prior to fix, the following would usually trigger the crash: chrome_frame_tests.exe --gtest_repeat=20 --gtest_filter=CF/FullTabUITest.TabCrashReload/0


Review URL: http://codereview.chromium.org/8965031

TBR=grt@chromium.org
Review URL: http://codereview.chromium.org/9005011