Change definition of "syncing" in AllStatus
This affects some stats displayed in about:sync.

Due to a logic error, syncing was defined as 'unsynced_handles > 0 &&
!throttled'.  This seems sort of reasonable, until you notice that the
unsynced_handles value is zero at the beginning of the sync cycle and is
not updated for successful commits that occur during the sync cycle.
And that's just the tip of the iceberg.

This change makes it so that 'syncing' is an edge-triggered value.  When
a sync cycle starts or ends, it will send a SYNC_CYCLE_BEGIN or
SYNC_CYCLE_END event to AllStatus, which will update the flag
accordingly.

This results in a different definition of 'syncing' than the previous
system.  Whereas the old system tried to infer what the SyncScheduler
was doing and remain in the syncing state until it had fully completed
its job, this one will report the 'syncing' state only if a call to
SyncShare() is currently in progress.  One of the implications of this
is that it will now toggle in and out of the syncing state when
committing more than 25 items, because the loop to commit all unsynced
items in batches of 25 currently occurs outside SyncShare.  The new
system will also bounce in an out of the syncing state as we handle the
SYNC_CYCLE_CONTINUATION cycle that follows every cycle that commits an
item.

How does this look in about:sync?  The biggest change is that it will
now be almost impossible to see the status summary of "SYNCING".  This
is because the ProfileSyncService does not receive notification of the
beginning of a sync cycle, only the end.  So it doesn't bother to update
the state of about:sync until the syncing is already completed.  Prior
to this change, about:sync did show the "SYNCING" summary sometimes, but
this was only during the *second* sync cycle, which we enter only because
of the SYNC_CYCLE_CONTINUATION bug (issue 94670).

Also in this commit are changes to remove the infrastructure used to
set the 'syncing' variable in the SyncSessionSnapshot.  This had the
side-effect of removing the mechanism by which we're guaranteed to kick
AllStatus with a STATUS_CHANGED event, so I replaced it with a
SYNC_CYCLE_BEGIN event sent from the SYNCER_BEGIN step.  I also removed
SyncerEndCommand, because the mechanims used to send the
SYNC_CYCLE_BEGIN event could be used to send the SYNC_CYCLE_ENDED
event, too.  

BUG=98346
TEST=


Review URL: http://codereview.chromium.org/8873058