Clean up VideoRendererBase timing code.
Tiny step towards simplifying VideoRendererBase code and respective subclasses.

The biggest change was to replace if statements adjusting the value of |remaining_time| with explicit statements to sleep then retry the loop from the start. The nice part about doing this is that we guarantee we exercise all conditions in the scenario where a spurious signal to |frame_available_| causes the thread to wake up.

BUG=28208

Review URL: http://codereview.chromium.org/8794001