Always send a reply to nacl launch requests.
Previously there were many return paths where the reply message would not get sent, which would indefinitely hang the renderer. I also found a leak where if the FILE thread is gone when we try to make the IRT available, the object will be leaked (the the reply message would not be sent either). I tried to find all of the failure cases and ensure the reply message is sent.

BUG=
TEST=

Review URL: http://codereview.chromium.org/8994010