This is the fourth patch in this series. This replaces the CHECK stmts with unrecoverable errors!! Syncable layer is now free of CHECK stmts. The only thing that could crash syncable layer(and hence the browser) would be actual memory corruption. DB corruptions should not crash the browser any more.
Note: We handle the return value only with in the syncable file. I.e., if a function returns error(false) we would unwind the stack without doing further processing inside the syncable layer. On the upper layers too we need to check the return value from syncable layer and unwind the stack in case of errors(The return value has to be propagated upto syncer or upto changeprocessor/Model Associator). Some of the plumbing is there and fair number of functions already check return value. Richard's CL would fix the rest. And I will pitch in if there are too many places.

The next cl would address restart after the unrecoverable error(It actually works today with no changes. But the code could use a little improvement) and along with it we will have a test that shutsdown pss, deletes the db and restarts the PSS.

BUG=100444
TEST=sync_integration_tests.exe, sync_unit_tests.exe, unit_tests.exe, manual tests 


Review URL: http://codereview.chromium.org/8897019