Add function support for Sleep with TimeDelta input.
This is the first step of fixing issue 108171 (converting calls of Sleep() to
use TimeDelta instead of ints in milliseconds).  I checked
platform_thread_unittests.cc for any tests explicitly for Sleep but found none,
so I didn't add any for this interface.

There will be a bit more implementation juggling here once the Sleep(int ms)
interface is removed, but that's coming in a later CL.

BUG=108171
TEST=


Review URL: http://codereview.chromium.org/8965072