chromeos: Get rid of a dependency to profile-guided optimization data.
The PGO data repository is as large as 116MB, thus not desirable
to pull unconditionally for developers.

Instead, we'll pull the dependency from chromeos-chrome ebuild file,
until we come up with a better solution to manage PGO data, not only
for Chrome OS, but also for other platforms.

BUG=108090
TEST=none. chrome builds without the data anyway.


Review URL: http://codereview.chromium.org/9113029