Make Panels to accept first mouse click w/o activation on Mac. This makes it consistent with mouse move events that are already dispatched into inactive always-on-top special windows like Balloons and Panels.
More background: on a Mac, the first click into a window usually is not dispatched into it, but rather is used to bring the window into foreground and 'activate' it. The special always-on-top windows like menus, balloons and panels do not need to be brought to the foreground so they usually do not require the intermediate 'activation' click. We used to pass the mouse moves but not mouse clicks which caused mouse hover effect without matching mouse click effect. See bug for more info.

BUG=105737


Review URL: http://codereview.chromium.org/9109025