Bind to eglCreateImageKHR and eglDestroyImageKHR when KHR_image_base is advertised by EGL.
At the moment we only bind to them if KHR_image (a superset) is advertised.

Because we determine to which extension each entry point belongs by parsing the header files and those entry points are not listed for KHR_image_base, I had to add some code to the autogenerator script to to add extra extensions for each entry point.

Review URL: http://codereview.chromium.org/9111039