In default X11 error handler, do not post task to message loop if there is no message loop.
There might not be a message loop if the X11 error happens early in initialization. This is the case for GLX initialization in the GPU process.

If there is no message loop, a simple error is logged containing the serial, error code, request code and minor code.

The purpose of posting a task to the message loop is to output a more descriptive error, involving calling various X11 functions such as XGetErrorText, which cannot be called in the error handler itself becasue of reentrancy issues.

BUG=109180
Review URL: http://codereview.chromium.org/9114014