Crash if the render process fails to create its IO thread.
We are seeing some latent crashes that I believe are a consequence of this earlier failure. So this change should help clarify what is happening.

A next step would be to determine the particular platform error that caused thread creation to fail.
(However base::PlatformThread doesn't currently expose that).

BUG=109302


Review URL: http://codereview.chromium.org/9122010