Move ChromotingHost initialization to the network thread.
Previously XmppSignalStrategy, ChromotingHost and some other objects were created on main or plugin thread but worked on the network thread. This could potentially cause synchronization bugs. Now it is possible to mark XmppSignalStrategy and LogToServer as NonThreadSafe to ensure that they are used only on the network thread.


Review URL: http://codereview.chromium.org/9051001