Add return values to SyncerCommand
This is part 1 of 3 in a series of patches to remove the unnecessary
SYNC_CYCLE_CONTINUATION sync cycle that follows every sync commit.

In this patch, we add a return value to all SyncerCommands.  This will
make it easier for us to tell when a command has failed, and why.

Currently, the error detection does not work very well.  We don't have
much support for errors that occur on worker threads.  Many commands will
return an error only in case of directory lookup failure, but return
NO_ERROR otherwise.

This is OK because we're not aiming to fully implement error detection
in this commit.  This provides a base we can build on in future commits.

BUG=94670,109422
TEST=


Review URL: http://codereview.chromium.org/9113024