Fix executable searching in keygen.py
Add the script's directory to the search path, so it can be used in a
standalone tarball for Virtual Me2Me.

Use os.path.join() instead of string concatenation.

Make the module safe to import (that is, don't search for remoting_host_keygen
whilst the module is being imported, because it fails to work from tools such
as pychecker etc).

BUG=None
TEST=Manual

Review URL: http://codereview.chromium.org/9114045