Remove PyAuto based event tests; covered by Layout Tests.
After some thought and discussion with Kenji, PyAuto is a pretty poor
place for this type of functional test.  Layout Tests are much better
suited for this type of testing.  In fact, there is already extensive
coverage of the media state machine in Layout Tests.  As such, delete
the PyAuto based event tests.

BUG=106263
TEST=none


Review URL: http://codereview.chromium.org/9158024