Make GLES2CmdHelper and CommandBufferHelper handle failed allocation
This passes all the current unit tests.
I'll write more tests to check this stuff but I want to get it the tree
asap to see if it stop the crashes

Note. This is only step 1 of about 4 steps
Step 2 is make GLES2Implementation do the same
Step 3 is make both of these trigger lost context
Step 4 is write tests.

TEST=none
BUG=109694

R=apatrick@chromium.org


Review URL: http://codereview.chromium.org/9166017